Share run_search_test functionality
